+++ src/files/grid_maps.cpp
@@ -129,8 +129,8 @@ void saveMessageMap(const char *imageFile,const GRID::Message *message,T::ParamV
     uint height = 900;
 
     uint size = (width*height);
-    unsigned long *image = new unsigned long[size];
-    memset(image,0xFF,size*sizeof(unsigned long));
+    uint *image = new uint[size];
+    memset(image,0xFF,size*sizeof(uint));
 
     unsigned char hue = 30;
     unsigned char saturation = 128;
